+++ components/organisms/o-my-account-shipping-details.vue
@@ -61,12 +61,22 @@
               :error-message="$t('Field is required.')"
               class="form__element form__element--half"
             />
-            <SfInput
+            <SfSelect
               v-model="editedAddress.state"
               name="state"
               :label="$t('State/Province')"
-              class="form__element form__element--half form__element--half-even"
-            />
+              required
+              :error-message="$t('Field is required.')"
+              class="sf-select--underlined form__select form__element form__element--half form__select form__element--half-even"
+            >
+              <SfSelectOption
+                v-for="state in stateOptions"
+                :key="state.code"
+                :value="state.name"
+              >
+                {{ state.name }}
+              </SfSelectOption>
+            </SfSelect>
             <SfInput
               v-model="editedAddress.postcode"
               name="postcode"
@@ -172,7 +182,9 @@ import pick from 'lodash-es/pick';
 import { required, minLength } from 'vuelidate/lib/validators';
 import { unicodeAlpha, unicodeAlphaNum } from '@vue-storefront/core/helpers/validators';
 import { SfTabs, SfInput, SfButton, SfSelect, SfIcon } from '@storefront-ui/vue';
+import config from 'config';
 const Countries = require('@vue-storefront/i18n/resource/countries.json')
+const States = require('@vue-storefront/i18n/resource/states.json')
 
 export default {
   name: 'OMyAccountShippingDetails',
@@ -192,12 +204,17 @@ export default {
         country: '',
         telephone: ''
       },
-      countries: Countries
+      countries: Countries,
+      states: States
     }
   },
   computed: {
     addresses () {
       return this.$store.state.user.current.addresses
+    },
+    stateOptions () {
+      let countryCode = this.editedAddress.country ? this.countries.find(country => country.name === this.editedAddress.country).code : config.i18n.defaultCountry
+      return this.states[countryCode] ? this.states[countryCode] : []
     }
   },
   methods: {
@@ -250,7 +267,8 @@ export default {
         lastname: this.editedAddress.lastname,
         street: [this.editedAddress.streetName, this.editedAddress.apartment],
         city: this.editedAddress.city,
-        ...(this.editedAddress.state ? { region: { region: this.editedAddress.state } } : {}),
+        ...(this.editedAddress.state && this.stateOptions ? { region: { region: this.editedAddress.state, region_id: this.stateOptions.find(state => state.name === this.editedAddress.state).code || this.editedAddress.state, region_code: this.stateOptions.find(state => state.name === this.editedAddress.state).code || this.editedAddress.state } } : {}),
+        ...(this.stateOptions ? { region_id: this.stateOptions.find(state => state.name === this.editedAddress.state).code || this.editedAddress.state } : {}),
         country_id: this.countries.find(country => country.name === this.editedAddress.country).code || this.editedAddress.country,
         postcode: this.editedAddress.postcode,
         ...(this.editedAddress.telephone ? { telephone: this.editedAddress.telephone } : {})
@@ -299,6 +317,9 @@ export default {
       postcode: {
         required,
         minLength: minLength(3)
+      },
+      state: {
+        required
       },
       country: {
         required