+++ CHANGELOG.md
@@ -89,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `isAddToCartDisabled` computed property (#377)
 - Update filters bar on category page (#381)
 - Use i18n wrapper for the login title (#438)
+- Added country dependent state dropdown (#431)
 - Corrected displayed the selected size option on product page (#436)
 - Fixed memory leak (#419)
 - Styling around color selector on filters in category page (#443)